requirements.txt
Flask==3.0.2
requests==2.31.0
python-dotenv==1.0.0
Flask-CORS==4.0.0
cryptography==42.0.7
pycryptodome==3.20.0
matplotlib==3.8.4
numpy==1.26.4
setuptools==70.0.0
wheel==0.43.0
Flask-Limiter==3.5.0
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
SQLAlchemy
marshmallow==3.20.1
redis==5.0.1
mysql-connector-python==8.1.0
PyJWT==2.8.0
